feat(button): remove typescript * update button * test customize button
