--- tsconfig.json
@@ -1,32 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "esnext",
-        "module": "esnext",
-        "moduleResolution": "node",
-        "lib": [
-            "es2017",
-            "dom"
-        ],
-        "jsx": "react",
-        "rootDirs": ["src", "stories"],
-        "baseUrl": "src",
-        "allowJs": true,
-        "checkJs": true,
-        "noEmit": true,
-        "strict": false,
-        "noImplicitThis": true,
-        "alwaysStrict": true,
-        "esModuleInterop": true,
-        "experimentalDecorators": true
-    },
-    "include": [
-        "packages/**/*",
-        "node_modules/@open-wc/**/*.js",
-        "d.ts"
-    ],
-    "exclude": [
-        "node_modules/!(@open-wc)",
-        "build",
-        "scripts"
-    ]
-}
