--- typings/declarations.d.ts
@@ -1,3 +1,0 @@
-// declare module '*.scss';
-// declare module '*.sass';
-// declare module '*.scss';
